minor style changes, minor change to pdf cleanup script, and new file names for nicer user interaction

+++ cleanup_timesheets.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
 files="$(find "$(dirname $0)/static/pdf" -iname "*.pdf" -mmin +30 | wc -l)"
-echo "[`date`]: Deleted $files generated pdf files"
 
-find "$(dirname $0)/static/pdf" -iname "*.pdf" -mmin +30 -delete
+if (( files > 0 )); then
+	echo "[`date`]: Deleted $files generated pdf files"
+
+	find "$(dirname $0)/static/pdf" -iname "*.pdf" -mmin +30 -delete
+fi
+
 
 # */30 * * * * /home/b21/timesheet/cleanup_timesheets.sh >> /home/b21/cleanup.log 2>&1
